refactor: Aded generic interfaces for storehandler and zarr_utils for support of zarr v2&v3
