+++ rkns/_zarr/storehandler_zarr_v2.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .storehandler_interface import _StoreHandler
+
+
+class StoreHandlerZarrV2(_StoreHandler):
+    pass
