+++ rkns/_zarr/utils_zarr_v2.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .utils_interface import ZarrUtils
+
+
+class _ZarrV2Utils(ZarrUtils):
+    pass
