+++ rkns/adapters/base.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from pathlib import Path
 
+from rkns._zarr import ZarrGroup
+from rkns._zarr.storehandler_zarr_v3 import _StoreHandler
 from rkns.file_formats import FileFormat
-from rkns.zarr.handler import StoreHandler
-from rkns.zarr.zarr_util import ZarrGroup
 
 __all__ = ["RKNSBaseAdapter"]
 
@@ -19,7 +19,7 @@ class RKNSBaseAdapter(ABC):
     Not technically an adapter in an object oriented sense, and more like a Data Converter.
     """
 
-    def __init__(self, handler: StoreHandler) -> None:
+    def __init__(self, handler: _StoreHandler) -> None:
         self._handler = handler
         super().__init__()
